requirements.txt
fastapi
uvicorn[standard]
pyyaml
regex
streamlit
pandas
pytest
ruff
mypy
python-dotenv
requests
